% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxapp_applicationcache.R
\docType{class}
\name{cxapp_applicationcache-class}
\alias{cxapp_applicationcache-class}
\alias{cxapp_applicationcache}
\title{Utility class to store and retrieve content in the application cache}
\description{
The application cache is a normalized file caching utility. An item is
referenced by a path syntax that maps to the item.

The location of the cache is set by the configuration option \code{CACHEPATH} in
\code{app.properties}. If \code{CACHEPATH} is not defined, the \code{.cache} sub-directory
in the temporary directory \link[base]{tempdir} for the current R session is
used.

The \code{add(x)} method will add the specified \code{file} to the cache using \code{x} as a
reference. The time to expire for the item in the cache is defined using
the configuration option \code{APPCACHE.EXPIRE} in minutes. If the expire duration
is updated, the duration in effect when the item was add or last touched is
used.

The \code{exists(x)} method returns a logical vector of the same length as \code{x}
indicating if the object reference exists in the cache (\code{TRUE}) or not
(\code{FALSE}). An object exists in the cache if it physically exists and the
object has not expired in the cache.

The \code{get(x)} method returns the path to the referenced object in the cache. Note
that the preference and the path is not the same. The returned path does not
preserve directory structure, file names or file extension.

The \code{drop(x)} method drops (deletes) an object from the cache.

The \code{touch(x)} method will reset the date and time for an object to expire in
the cache. The new date and time of expiration is derived from the current
value of the configuration option \code{APPCACHE.EXPIRE}.

The \code{purge()} method deletes all objects in the cache.
}
\section{Methods}{

\describe{
\item{\code{add(x)}}{Add one or more files to cache}

\item{\code{drop(x)}}{Drop an item from cache}

\item{\code{exists(x)}}{Check if item exists in cache}

\item{\code{get(x)}}{Get the path to the referenced item in cache}

\item{\code{initialize()}}{Initialize}

\item{\code{purge()}}{Delete all items in cache}

\item{\code{touch(x)}}{Extend expiration for an item in cache}
}}

