% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxapp_authapi.R
\name{cxapp_authapi}
\alias{cxapp_authapi}
\title{Utility function to authenticate an  API authorization request}
\usage{
cxapp_authapi(x)
}
\arguments{
\item{x}{Authorization string}
}
\value{
\code{TRUE} if authorization is valid. \code{FALSE}, otherwise
}
\description{
The function currently only supports bearer method of authorization. If the
authorization is not a bearer method, the result is surely \code{FALSE}.

The function also only supports bearer tokens registered as a secret in a
supported key vault (see \link[cxapp]{cxapp_vault}).

The space delimited list of named secrets is defined using the configuration option
\code{API.AUTH.SECRETS} in the app.properties file (see \link[cxapp]{cxapp_config}).
If no secrets are defined, the authorization result is surely \code{FALSE}.

The function searches the defined list of secrets for a match to the encoded
bearer string, excluding the \emph{bearer} prefix. The search assumes the secret
is a JSON data structure with a named element \code{value} retaining the encoded
string.

If a secret ends in an asterisk \code{*}, the secret name is assumed to be a prefix
and all secrets with the prefix is searched.

The use of the hierarchical reference in secret names is supported. See the
respective vault configuration.

If the encoded bearer string matches the secret value, the function returns
\code{TRUE}.

The result attribute \code{scope}, if defined, denotes if the authorization applies
to a service or particular user.

The result attribute \code{principal}, if defined, names the service or user
associated with the authorization.
}
