% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxapp_datapath.R
\name{cxapp_datapath}
\alias{cxapp_datapath}
\title{Utility function to get the path in the data area}
\usage{
cxapp_datapath(...)
}
\arguments{
\item{...}{Vector of path elements}
}
\value{
A vector of length one with the path
}
\description{
The utility function \code{cxapp_datapath()} is synonymous with the function
\link[base]{file.path}.

The root of the data path is defined in the cxapp property \code{DATA}. The first
path in \code{DATA} that exists is the root of the returned path. If no specified
paths exists, the first specified path is used.

If the \code{DATA} property is not defined, the function returns the path equal to
that of \link[base]{file.path}.
}
