% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxapp_apitokenencode.R
\name{.cxapp_apitokenencode}
\alias{.cxapp_apitokenencode}
\title{Internal utility to encode an API token}
\usage{
.cxapp_apitokenencode(x)
}
\arguments{
\item{x}{Token as a string}
}
\value{
Character vector of length 1 containing the encoded token
}
\description{
The utility function is a one-way encoding of a token string. The encoding
does not use a salt but rellies on a hashing strategy with information loss.
}
\keyword{internal}
