% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxapp_standardpath.R
\name{.cxapp_standardpath}
\alias{.cxapp_standardpath}
\title{Internal utility to standardise path separators in path references}
\usage{
.cxapp_standardpath(x)
}
\arguments{
\item{x}{A vector of paths}
}
\value{
A vector of standardized paths
}
\description{
A path uses separators of directories and file references. On Microsoft
Windows this is a backslash (\verb{\\}). On Linux the separator is a forward slash.

In addtion, the use of the backward slash will also need to be escaped as the
backward slash is an escape character.

R supports the foward slash on Windows environments, however not all functions
allows the path separator to be specified.

The function will translate all backward slashes into a forward slash.
}
\keyword{internal}
