% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxappconfig.R
\name{.cxappconfig}
\alias{.cxappconfig}
\title{Utility function to implement a global config object}
\usage{
.cxappconfig(x)
}
\arguments{
\item{x}{Configuration object}
}
\value{
Configuration object
}
\description{
Configurations are used extensively throughout the cxapp package and importing
configurations from property files every time a configuration is required will
include unnecessary processing for a very static reference.

The function will return the \emph{cached} \link[cxapp]{cxapp_config}
object from the global environment.

The configuration object is \code{.cxapp.wrkcache.appconfig} in the
\link[base]{.GlobalEnv} environment.

If a configuration object is specified as input \code{x}, the global configuration
is updated before it is returned.
}
