% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxapp_config.R
\docType{class}
\name{cxapp_config-class}
\alias{cxapp_config-class}
\alias{cxapp_config}
\title{Utility class to represent app configurations}
\description{
A utility class to represent app configurations defined in property files and
environmental variables.

Property files are named \verb{<context>.properties} where the configuration option
is referred to by \verb{context/<property>}.

Property file syntax and conventions are specified in the help reference for
function \link[cxapp]{cxapp_propertiesread}.

Class initialization takes a vector of paths as an argument. If the path
specified ends in \code{.properties}, it is assumed a properties file. Otherwise,
the entry is assumed a directory containing property files. The vector \code{x}
is processed in specified order and files within a directory in natural sort
order.

A property file name contains the characters a-z and digits 0-9 and the file
extension \code{properties}. The property file name excluding the file extension
is used as the context to look up a named property value.

The \code{option()} method returns the value of an option if it exists or the value
of \code{unset} if the option does not exist. An option is referred to by the string
\verb{<context>/<property>}. If the context is specified, the context is assumed
to be \code{cxapp}.

If the option is not defined as part of a property file, the \code{option()} method
searches for an environment variable \verb{<context>_<property>}. Any periods in
property name part is converted to underscores.

An option value that contains the prefix \verb{[env] <name>} or starts with the
character \verb{$<name>} is interpreted as a reference to an environmental
variable with specified name. If the specified environmental variable is not
defined, the value of \code{unset} is returned. The environmental variable name is
case sensitive with leading and trailing spaces removed.

An option value that contains the prefix \verb{[vault] <name>} is interpreted as
a reference to a vault secret with specified name. If a vault service is not
configured or available or the vault hs not defined the specified secret, the
value of \code{unset} is returned. The secret name is case sensitive with leading
and trailing spaces removed.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal configuration data store}
}}

\section{Methods}{

\describe{
\item{\code{initialize(x)}}{Initialize}

\item{\code{option(x, unset = NA, as.type = TRUE)}}{Get property value}
}}

