% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxapp_log.R
\name{cxapp_log}
\alias{cxapp_log}
\title{Simple function to log messages}
\usage{
cxapp_log(x, attr = NULL, echo = base::interactive())
}
\arguments{
\item{x}{Vector of messages to log}

\item{attr}{A vector of attribute values to append to log message}

\item{echo}{Enable/Disable output of log messages to console}
}
\value{
Invisible vector of log messages
}
\description{
The \code{cxapp_log} function logs messages in specified order prefixed with the current
date and time.

The log file parent directory path is defined by the \code{LOG.PATH} cxapp property. If
\code{LOG.PATH} is not defined log messages are written to the console.

The \code{LOG.NAME} option specifies the basis for the log file name. Any log file
name parts associated with log rotation is appended to the log file base name
separated by an underscore (\verb{_}). Log file names end in the file extension \code{log}.

If \code{LOG.NAME} is not defined, the log file name is \code{app.log}.

The \code{LOG.ROTATION} cxapp property defines the log file rotation. Valid rotations
are \code{YEAR}, \code{MONTH} and \code{DAY}. The log rotation follows the format four digit
year and two digit month and day.

Log messages follow the convention \verb{[\%Y-\%m-\%d \%H:\%M:\%S] <message>}

Specified attribute values are appended to the log message between square
brackets delimited by semi-colon.

if executing in \code{interactive()} mode, log messages are written to the console.
}
